Search_Agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Implementation of simple search agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.11.0",
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
]
